# Conflicts:
#	src/main/java/com/spring/security/medi/care/app/afiliacion/service/SolicitudAfiliacionServiceImpl.java
#	src/main/java/com/spring/security/medi/care/app/commons/domain/Ciudadano.java
#	src/main/java/com/spring/security/medi/care/app/controller/dto/SolicitudAfiliacionOutputDto.java
#	src/main/java/com/spring/security/medi/care/app/controller/rest/UsuarioRestController.java
#	src/main/java/com/spring/security/medi/care/app/usuario/service/UsuarioService.java
#	src/main/java/com/spring/security/medi/care/app/usuario/service/UsuarioServiceImpl.java
#	src/main/resources/templates/pages/despacho/show.html

+++ src/main/java/com/spring/security/medi/care/app/controller/dto/ErrorPageDto.java
@@ -11,11 +11,11 @@ public class ErrorPageDto implements Serializable {
 
     private String descripcionError;
 
-    private Integer codigoError;
+    private Long codigoError;
 
     private boolean isError = false;
 
-    public ErrorPageDto(Integer codigoError, String descripcionError, boolean isError) {
+    public ErrorPageDto(Long codigoError, String descripcionError, boolean isError) {
         this.descripcionError = descripcionError;
         this.codigoError = codigoError;
         this.isError = isError;
